refactor: split VoterAuthentication into molecules
